backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
bcrypt==4.0.1
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
